make interaction interpoint types string
