+++ plugins/optimizer/coordinator/tasks.py
@@ -19,13 +19,13 @@ import requests
 from celery.utils.log import get_task_logger
 
 from plugins.optimizer.coordinator import Optimizer
+from plugins.optimizer.shared.enums import InteractionEndpointType
 from plugins.optimizer.shared.schemas import (
     ObjectiveFunctionPassData,
     ObjectiveFunctionPassDataResponse,
     ObjectiveFunctionPassDataResponseSchema,
     ObjectiveFunctionPassDataSchema,
 )
-from qhana_plugin_runner.api.plugin_schemas import InteractionEndpointType
 from qhana_plugin_runner.api.tasks_api import TaskData, TaskStatusSchema
 from qhana_plugin_runner.celery import CELERY
 from qhana_plugin_runner.db.models.tasks import ProcessingTask
